+++ packages/terra-application-links/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Updated to interface with react-intl's Intl context rather than old React Context API
 
 6.5.0 - (July 10, 2019)
 ------------------
@@ -13,7 +15,6 @@ Unreleased
 ------------------
 ### Changed
 * Renamed test and example SCSS files to end with `.module.scss` extension
-* Updated to interface with react-intl's Intl context rather than old React Context API
 
 6.3.0 - (June 19, 2019)
 ------------------