+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Updated to interface with react-intl's Intl context rather than old React Context API
 
 4.6.0 - (July 10, 2019)
 ------------------
@@ -18,7 +20,6 @@ Unreleased
 ### Changed
 * Removed extra inline styles.
 * Renamed test and example SCSS files to end with `.module.scss` extension
-* Updated to interface with react-intl's Intl context rather than old React Context API
 
 4.4.0 - (June 19, 2019)
 ------------------